requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0
django-dotenv==1.4.2
django-jazzmin==3.0.1
django-querycount==0.8.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
swagger-spec-validator==3.0.4
typing_extensions==4.15.0
uritemplate==4.2.0
